refactor: extract daemon orchestration into dedicated modules (#160)
Restore src/daemon.ts to thin-router entrypoint by extracting:
- server-lifecycle.ts: lock/info file management, code signature, env parsing
- request-router.ts: request dispatch orchestration across handlers
- transport.ts: socket server, listen helpers, disconnect/abort logic
